+++ internal/app/types/stores.go
@@ -25,6 +25,7 @@ type UsersStore interface {
 type WebAppsStore interface {
 	dbStore
 	CountByFilter(ftr Filter) (int, error)
+	DeleteByIdAndUserEmail(id, usr string) error
 	GetByFilter(ftr Filter, pag Pagination) ([]*WebApp, error)
 	GetByIdAndUserEmail(id, usr string) (*WebApp, error)
 	Insert(wap *WebApp) (*WebApp, error)